Rename to `_assertions.py`
