+++ tests/http/test_web_server.py
@@ -27,7 +27,7 @@ from proxy.common.utils import build_http_response, build_http_request, bytes_
 from proxy.common.constants import CRLF, PLUGIN_HTTP_PROXY, PLUGIN_PAC_FILE, PLUGIN_WEB_SERVER, PROXY_PY_DIR
 from proxy.http.server import HttpWebServerPlugin
 
-from ..assertions import Assertions
+from .._assertions import Assertions
 
 
 PAC_FILE_PATH = os.path.join(
