+++ tests/plugin/test_http_proxy_plugins_with_tls_interception.py
@@ -26,7 +26,7 @@ from proxy.http.parser import HttpParser
 
 from .utils import get_plugin_by_test_name
 
-from ..assertions import Assertions
+from .._assertions import Assertions
 
 
 class TestHttpProxyPluginExamplesWithTlsInterception(Assertions):
